requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# circle_dataset.py: 6
# main.py: 7
# train.py: 14
Pillow == 6.1.0

# generate_data.py: 2
# main.py: 2
Shapely == 1.6.4.post2

# generate_data.py: 8
matplotlib == 3.0.3

# circle_dataset.py: 7
# generate_data.py: 1
# main.py: 1
# train.py: 12
numpy == 1.17.2

# circle_dataset.py: 4
pandas == 0.22.0

# circle_dataset.py: 5
# generate_data.py: 3
# main.py: 3
scikit_image == 0.15.0

# generate_data.py: 4
# main.py: 8
scipy == 1.1.0

# circle_dataset.py: 3,8
# main.py: 5,6
# models/resnet.py: 2
# train.py: 1,3,4,5,6
torch == 1.2.0

# main.py: 10
# train.py: 7
torchvision == 0.4.0
